# ⚡️(api) improve performance of HTTP basic auth