--- a/docs/api.md
+++ b/docs/api.md
@@ -154,3 +154,20 @@ If you want to lower noisy transactions (_e.g._ in a Kubernetes cluster), you ca
 ```bash
 RALPH_SENTRY_IGNORE_HEALTH_CHECKS=True
 ```
+
+### Additional configuration
+
+#### HTTP Basic auth caching 
+
+HTTP basic auth implementation uses the secure and standard bcrypt algorithm to hash/salt passwords before storing them.
+This implementation comes with a performance cost.
+To speed up requests, credentials are stored in a LRU cache with a Time To Live.
+To configure this cache, you can define the following environment variables:
+- the maximum number of entries in the cache. Select a value greater than the maximum number of individual user credentials, for better performance. Defaults to 100. 
+```
+RALPH_AUTH_CACHE_MAX_SIZE=100
+```
+- the Time To Live of the cache entries in seconds. Defaults to 3600s.
+```
+RALPH_AUTH_CACHE_TTL=3600
+```