--- a/setup.cfg
+++ b/setup.cfg
@@ -87,6 +87,7 @@ ci =
 lrs =
     bcrypt==4.0.1
     fastapi==0.96.0
+    cachetools==5.3.0
     ; We temporary pin `h11` to avoid pip downloading the latest version to solve a
     ; dependency conflict caused by `httpx` which requires httpcore>=0.15.0,<0.16.0 and
     ; `httpcore` depends on h11>=0.11,<0.13.