@wilbrdt 2023-05-30
> Maybe you should also add python tests ensuring we hit the cache on subsequent requests?

We would either need to do a timing test, or to activate with `info=True`  instrumentation that gives us access to `cache_info()`, but that would induce a performance penalty for each requests.